requirements-dev.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
astroid==2.6.2; python_version ~= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
azure-functions==1.7.2
backports.entry-points-selectable==1.1.0; python_version >= '2.7'
certifi==2021.5.30
cfgv==3.3.0; python_full_version >= '3.6.1'
charset-normalizer==2.0.1; python_version >= '3'
coverage==5.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
coveralls==3.1.0
distlib==0.3.2
docopt==0.6.2
filelock==3.0.12
identify==2.2.11; python_full_version >= '3.6.1'
idna==3.2; python_version >= '3'
importlib-metadata==4.6.1; python_version < '3.8'
iniconfig==1.1.1
isort==5.9.2; python_version < '4' and python_full_version >= '3.6.1'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.0; python_version >= '3.6'
platformdirs==2.0.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pre-commit==2.13.0
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint==2.9.3
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-cov==2.12.1
pytest==6.2.4
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typed-ast==1.4.3; python_version < '3.8' and implementation_name == 'cpython'
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
virtualenv==20.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.12.1
zipp==3.5.0; python_version >= '3.6'
